requirements.txt
numpy
tensorflow==2.1.0
Pillow
